feat: use local bin if found
