+++ packages/umi/package.json
@@ -9,10 +9,12 @@
     "babel-preset-umi": "^1.0.0-beta.3",
     "chai": "^4.1.2",
     "cross-spawn": "^6.0.5",
+    "debug": "^3.1.0",
     "ejs": "^2.5.7",
     "is-windows": "^1.0.2",
     "path-is-absolute": "^1.0.1",
     "query-string": "^5.0.0",
+    "resolve-cwd": "^2.0.0",
     "slash2": "^2.0.0",
     "umi-build-dev": "^1.0.0-beta.5",
     "update-notifier": "^2.3.0",
